Split messages/src/lib.rs
